Update capitalizations + check links (#70)
* Update capitalizations + check links 

All links are correct. Remove 'encryption' heading in table of contents b/c it is redundant. Update case of headings / subheadings to be consistent. The term 'enclave' is consistently lower case in other parts of the documentation so made document reflect that.

* Update missed capitalization

Fix capitalization issues missed from first pass through.