Update quickstart code + copy  (#92)
* Remove publish

Signed-off-by: Dylan Schultz <dylanschultz311@gmail.com>

* Remove dead links

Signed-off-by: Dylan Schultz <dylanschultz311@gmail.com>

* Update Secret Node version

Signed-off-by: Dylan Schultz <dylanschultz311@gmail.com>

* Add correct shasum

* Remove quicksync reference

* Fix outdated code on docker code example

secretcli is now secretd -- this must be used when interacting with files inside of the docker container.

* Fix Querying the Smart Contract and Code

Update secretcli to secretd, and add additional code comments.

* Update Instantiate the Smart Contract code 

Update secretcli to secretd. All code in this section is working as expected.

* Update testnet node address and config chain-id 

The testnet node is now updated to correct testnet node link, and the config chain-id is now updated to correct testnet chain-id --> pulsar-2

* Undo last commit changes

Need to work on getting connected to the testnet further.

* Update test net code 

chain-id and config node were outdated. They now contain working links to the rpc and correct test net chain-id

* Update querying account balance code 

Now there is a dedicated section to confirm that the faucet funded the correct address that a user will be deploying their Secret Contract with to the test net.

* Add show address code 

Added section dedicated to viewing Secret address using the CLI.

* Correct code 

Forgot to add 'secretcli' on previous code example

* Fix code snippet showing address

* Update code for uploading contract to test net 

added gas and gas-prices flags to code to successfully deploy contract to the test net.

* Update code / text to Pulsar-2

Reference to Holodeck removed and replaced with pulsar-2

* Check links + final copy update

All links are working as aspected. Finished updating copy.

* Update secretd + add pulsar-2

Updated 'secretcli' to 'secretd' and updated holodeck references to pulsar-2.

* Copy update

* Minor copy update

Co-authored-by: Dylan Schultz <dylanschultz311@gmail.com>
Co-authored-by: moonstash <15162813+moonstash@users.noreply.github.com>